make it able to comment from all pages
