+++ src/js/feed.js
@@ -9,12 +9,12 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 import endpointObject from "./endpoints.js";
 import callApi from "./callApi.js";
-import renderPosts, { renderComments } from "./renderPost.js";
+import renderPosts from "./renderPost.js";
 import filterPosts from "./filter.js";
 import createSmileyPicker from "./emoji.js";
 import reactToPost from "./reactToPost.js";
-import commentPost from "./commentPost.js";
 import fadeText from "./fadeText.js";
+import commentButton from "./commentOnClick.js";
 const endpoint = endpointObject("Jarle");
 const [setSmiley, setId, getSmiley, getId] = createSmileyPicker();
 let modal = document.querySelector("#modal");
@@ -228,18 +228,6 @@ function emojiReactButton() {
             console.log(getId());
         });
     });
-}
-function commentButton() {
-    document.querySelectorAll("[data-comment-id]").forEach((button) => button.addEventListener("click", () => {
-        const id = button.dataset.commentId;
-        console.log(id);
-        const message = document.querySelector(`#commentInput${id}`).value;
-        commentPost(message, id);
-        renderComments(document.querySelector(`#comment-row--${id}`), message, "", {
-            name: JSON.parse(localStorage.getItem("currentUser")),
-            avatar: JSON.parse(localStorage.getItem("avatar")),
-        });
-    }));
 }
 (function renderUserSpecific() {
     var _a;
