+++ src/js/feed.ts
@@ -6,6 +6,7 @@ import createSmileyPicker from "./emoji.js";
 import reactToPost from "./reactToPost.js";
 import commentPost from "./commentPost.js";
 import fadeText from "./fadeText.js";
+import commentButton from "./commentOnClick.js";
 
 const endpoint = endpointObject("Jarle");
 const [setSmiley, setId, getSmiley, getId] = createSmileyPicker();
@@ -326,26 +327,6 @@ function emojiReactButton() {
       console.log(getId());
     });
   });
-}
-
-function commentButton() {
-  document.querySelectorAll("[data-comment-id]").forEach((button) =>
-    button.addEventListener("click", () => {
-      const id = button.dataset.commentId;
-      console.log(id);
-      const message = document.querySelector(`#commentInput${id}`).value;
-      commentPost(message, id);
-      renderComments(
-        document.querySelector(`#comment-row--${id}`),
-        message,
-        "",
-        {
-          name: JSON.parse(localStorage.getItem("currentUser")),
-          avatar: JSON.parse(localStorage.getItem("avatar")),
-        }
-      );
-    })
-  );
 }
 
 (function renderUserSpecific() {
