+++ src/js/singlePost.js
@@ -11,6 +11,7 @@ import callApi from "./callApi.js";
 import endpointObject from "./endpoints.js";
 import optionFactory from "./optionFactory.js";
 import renderPosts from "./renderPost.js";
+import commentButton from "./commentOnClick.js";
 const params = new URLSearchParams(window.location.search);
 const id = params.get("id");
 const endpoint = endpointObject("Jarle");
@@ -19,4 +20,5 @@ const getId = optionFactory("GET", {}, endpoint);
     const data = yield callApi(endpoint.getId(id), getId);
     console.log("data:", data);
     renderPosts(document.querySelector("#postContainer"), data);
+    commentButton();
 }))();
