+++ src/js/singlePost.ts
@@ -2,6 +2,7 @@ import callApi from "./callApi.js";
 import endpointObject from "./endpoints.js";
 import optionFactory from "./optionFactory.js";
 import renderPosts from "./renderPost.js";
+import commentButton from "./commentOnClick.js";
 const params = new URLSearchParams(window.location.search);
 const id = params.get("id");
 const endpoint = endpointObject("Jarle");
@@ -11,4 +12,5 @@ const getId = optionFactory("GET", {}, endpoint);
   const data = await callApi(endpoint.getId(id), getId);
   console.log("data:", data);
   renderPosts(document.querySelector("#postContainer"), data);
+  commentButton();
 })();
